// CodeApps/EventsHub/src/generated/models/Crb5a_codeapps_eventdetailsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface crb5a_codeapps_eventdetailses {
  crb5a_codeapps_eventdetailsid: string;
  crb5a_enddate: Date;
  crb5a_eventtype: string;
  crb5a_eventtypename?: string;
  crb5a_locationdetails?: string;
  crb5a_maindescription?: string;
  crb5a_shortdescription: string;
  crb5a_startdate: Date;
  crb5a_title: string;
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}